+++ server.js
@@ -12,7 +12,7 @@ const MongoStore = require('connect-mongo')(session);
 const mongoose = require("mongoose");
 
 const passport = require('./passport/setup');
-const auth = require('./routes/auth');
+// const auth = require('./routes/auth');
 
 const routes = require("./routes");
 
@@ -34,7 +34,7 @@ mongoose
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-
+// Express Session
 app.use(
   session({
     secret: 'super duper ultra secret',
@@ -49,18 +49,17 @@ if (process.env.NODE_ENV === "production") {
   app.use(express.static("client/build"));
 }
 
-app.use(function(req, res, next) {
-  res.header("Access-Control-Allow-Origin", "*");
-  res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
-  next();
-});
+// app.use(function(req, res, next) {
+//   res.header("Access-Control-Allow-Origin", "*");
+//   res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
+//   next();
+// });
 // Add routes, both API and view
-app.use(routes);
-
 app.use(passport.initialize());
 app.use(passport.session());
 
-app.use('/api/auth', auth);
+app.use(routes);
+
 
 // Start the API server
 app.listen(PORT, function() {
